+++ CodeGen/RuntimeSerializationSettingsUtils.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.IO;
 using UnityEngine;
@@ -7,29 +6,34 @@ namespace Unity.RuntimeSceneSerialization.CodeGen
 {
     static class RuntimeSerializationSettingsUtils
     {
-        const string k_AssemblyExceptionsPath = "ProjectSettings/RuntimeSerializationAssemblyExceptions.txt";
+        const string k_AssemblyInclusionsPath = "ProjectSettings/RuntimeSerializationAssemblyInclusions.txt";
         const string k_NamespaceExceptionsPath = "ProjectSettings/RuntimeSerializationNamespaceExceptions.txt";
         const string k_TypeExceptionsPath = "ProjectSettings/RuntimeSerializationTypeExceptions.txt";
 
-        static HashSet<string> s_AssemblyExceptions;
+        static HashSet<string> s_AssemblyInclusions;
         static HashSet<string> s_NamespaceExceptions;
         static HashSet<string> s_TypeExceptions;
 
         // Local method use only -- created here to reduce garbage collection. Collections must be cleared before use
         static readonly List<string> k_SortableList = new List<string>();
 
-        internal static void ClearExceptionCache()
+        internal static void ClearSettingsCache()
         {
-            s_AssemblyExceptions = null;
+            s_AssemblyInclusions = null;
             s_NamespaceExceptions = null;
             s_TypeExceptions = null;
         }
 
-        internal static HashSet<string> GetAssemblyExceptions()
+        internal static HashSet<string> GetAssemblyInclusions()
         {
-            return s_AssemblyExceptions ?? (s_AssemblyExceptions = File.Exists(k_AssemblyExceptionsPath)
-                ? new HashSet<string>(File.ReadAllLines(k_AssemblyExceptionsPath))
-                : new HashSet<string>());
+            if (s_AssemblyInclusions != null)
+                return s_AssemblyInclusions;
+
+            s_AssemblyInclusions = File.Exists(k_AssemblyInclusionsPath)
+                ? new HashSet<string>(File.ReadAllLines(k_AssemblyInclusionsPath))
+                : new HashSet<string>();
+
+            return s_AssemblyInclusions;
         }
 
         internal static HashSet<string> GetNamespaceExceptions()
@@ -48,7 +52,7 @@ namespace Unity.RuntimeSceneSerialization.CodeGen
 
         internal static void SaveSettings()
         {
-            SaveExceptions(k_AssemblyExceptionsPath, s_AssemblyExceptions);
+            SaveExceptions(k_AssemblyInclusionsPath, s_AssemblyInclusions);
             SaveExceptions(k_NamespaceExceptionsPath, s_NamespaceExceptions);
             SaveExceptions(k_TypeExceptionsPath, s_TypeExceptions);
         }